--- a/src/RunWrappers/Docker.js
+++ b/src/RunWrappers/Docker.js
@@ -131,7 +131,7 @@ Runner.prototype.matchWithoutVersion = function(a, b) {
 };
 
 Runner.prototype.containerNameMatches = function(container, name) {
-  return _.any(container.Names, function(val) {
+  return _.some(container.Names, function(val) {
     return val === '/' + name;
   });
 };