@cliftonc 2016-08-11
/cc @geophree - as discussed I was busy blowing my laptop up this morning.