@geophree 2016-08-11
Don't decrease the coverage!  :frowning: