@cliftonc 2016-08-12
Given we only have tests for src, adding any lines outside of src reduces coverage :(