@geophree 2016-08-12
I'm still a little bit worried about this with people on the new docker for mac.  You said it got better for you after an update?