@cliftonc 2016-08-12
It's fine for me, I think I had other issues with my laptop this morning.

On Thu, 11 Aug 2016, 19:29 Jeff Hanke, <notifications@github.com> wrote:

> I'm still a little bit worried about this with people on the new docker
> for mac. You said it got better for you after an update?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/tes/bosco/pull/163#issuecomment-239347304>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAOmybVQsavgOeVqzVtX6P_g1QqxpfW3ks5qe9qjgaJpZM4Jilu3>
> .
>