@cliftonc 2016-08-18
This would be a good thing to merge, can you finish your review @geophree?  If not can someone else take a look?  @dmorgantini @tomruttle @bettiolo?