@dmorgantini 2016-08-22
I'm gonna merge.  Semantic discussions can continue post-merge.